requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
django-admin==2.0.1
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-fsm==2.7.1
django-phone-field==1.8.1
django-six==1.0.4
gunicorn==20.0.4
Pillow==8.1.0
pytz==2020.5
screen==1.0.1
soupsieve==2.1
sqlparse==0.4.1
whitenoise==5.2.0
xlwt==1.3.0
